# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Snaptext Developer API",
    "version": "1.0.0",
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Format: Bearer st-<key>"
      }
    },
    "schemas": {
      "CancelJobResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the job"
          },
          "runId": {
            "type": "string",
            "description": "Underlying workflow execution run ID"
          },
          "status": {
            "type": "string",
            "enum": [
              "cancelled"
            ],
            "description": "Current status after cancellation"
          }
        },
        "required": [
          "jobId",
          "runId",
          "status"
        ]
      },
...
      "JobDetailsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "runId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "totalPages": {
                "type": "number"
              },
              "completedPages": {
                "type": "number"
              },
              "pdfUrl": {
                "type": "string",
                "format": "uri"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "hasSchema": {
                "type": "boolean"
              },
              "schema": {
                "type": "string",
                "nullable": true
              },
              "modelName": {
                "type": "string",
                "nullable": true
              },
              "metrics": {
                "type": "object",
                "properties": {
                  "extractedPages": {
                    "type": "number"
                  },
                  "emptyPages": {
                    "type": "number"
                  },
                  "failedPages": {
                    "type": "number"
                  },
                  "mergeablePages": {
                    "type": "number"
                  }
                },
                "required": [
                  "extractedPages",
                  "emptyPages",
                  "failedPages",
                  "mergeablePages"
                ]
              }
            },
            "required": [
              "totalPages",
              "completedPages",
              "pdfUrl",
              "createdAt",
              "updatedAt",
              "hasSchema",
              "schema",
              "modelName",
              "metrics"
            ]
          },
          "data": {
            "nullable": true,
            "description": "Merged layout extraction structure of the document (JSON shape)"
          },
          "pagesData": {
            "type": "array",
            "items": {},
            "description": "Individual extracted layout page outputs"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "runId",
          "status",
          "filename",
          "metadata",
          "data",
          "pagesData",
          "error"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/jobs/{id}": {
      "get": {
        "operationId": "get-v1-jobs-{id}",
        "summary": "Get Job Status & Results",
        "description": "Retrieves real-time processing status, metadata metrics, and final merged schema layouts for a specific job.",
        "tags": [
          "V1"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the job or wrun_ workflow execution run ID"
            },
            "description": "UUID of the job or wrun_ workflow execution run ID",
            "example": "123"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/v1/jobs/{id}/cancel": {
      "post": {
        "operationId": "post-v1-jobs-{id}-cancel",
        "summary": "Cancel OCR Job",
        "description": "Gracefully halts and aborts an active document processing workflow run.",
        "tags": [
          "V1"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the job or wrun_ workflow execution run ID"
            },
            "description": "UUID of the job or wrun_ workflow execution run ID",
            "example": "123"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelJobResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
...
}